convert_exl2: long context quantization
Tweaks have added to quantize for use with longer context and
focus on coding c++ and python.
Added option -cdf <num> for scaling up or down the amout of
code and technical data samples for measurement and calibration.
Removed limit and fixed issues with longer that 2048 input
tokens for measurement and calibration.
Added scripts for generating more calibration data and
running different quantizations on different GPUs in parallel
(not distributed, just runs each quant on its own GPU).